+++ allauth/socialaccount/models.py
@@ -297,7 +297,7 @@ class SocialLogin(object):
     @classmethod
     def stash_state(cls, request):
         state = cls.state_from_request(request)
-        verifier = get_random_string()
+        verifier = get_random_string(12)
         request.session['socialaccount_state'] = (state, verifier)
         return verifier